+++ packages/core/api/src/ping/ping.controller.ts
@@ -1,0 +1,11 @@
+import { Controller, Get } from '@nestjs/common';
+import { Public } from '@/core';
+
+@Controller('ping')
+export class PingController {
+  @Public()
+  @Get()
+  async ping() {
+    return { ts: Date.now() };
+  }
+}