+++ packages/core/web/src/core/repository/core.repository.ts
@@ -1,6 +1,8 @@
 import axios from 'axios';
 import { Icons } from '@lyvely/ui';
 import { useGlobalDialogStore } from '../store';
+import { useFlashStore, usePageStore } from '@/ui';
+import { usePingService } from '@/ping';
 
 // TODO: abstract this away in config or something..
 const apiURL = import.meta.env.VITE_APP_API_URL || 'http://localhost:8080';
@@ -25,15 +27,34 @@ export function createFileUrl(hash: string) {
 
 repository.defaults.withCredentials = true;
 
+let pingInterval: number | undefined = undefined;
+
 repository.interceptors.response.use(undefined, (error) => {
   return new Promise((resolve, reject) => {
     if (!error.response) {
-      useGlobalDialogStore().showError({
-        icon: Icons.error_network.name,
-        title: 'error.network.title',
-        message: 'error.network.message',
-        buttonType: 'reload',
-      });
+      if (usePageStore().loaded) {
+        useFlashStore().addErrorFlash('error.network.message');
+      } else {
+        useGlobalDialogStore().showError({
+          icon: Icons.error_network.name,
+          title: 'error.network.title',
+          message: 'error.network.message',
+          buttonType: 'reload',
+        });
+      }
+
+      if (!pingInterval) {
+        pingInterval = setInterval(() => {
+          usePingService()
+            .ping()
+            .then((result) => {
+              if (result) {
+                clearInterval(pingInterval);
+                document.location.reload();
+              }
+            });
+        }, 2000) as any;
+      }
     } else if (
       error.response.status === 403 &&
       error.response.data.message === 'invalid csrf token'