+++ packages/core/web/src/ping/ping.repository.ts
@@ -1,0 +1,7 @@
+import { repository } from '@/core';
+
+export default {
+  async ping() {
+    return repository.get<{ ts: number }>('ping');
+  },
+};